Move `RemoteFeedImageDataLoader` to production in its own file.
